Resolve logout redirect
after the query stopped retrying, 'isFetching' would be 'false' but
there was no new 'profile' value to replace the stale value. so the
condition i was using before was still "true".

To ensure that didn't happen, it was necessary to also check 'isStale'

I'm  thinking it'd be nicer to just disabled this query after the first
time it resolves tho..